requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
future==1.0.0
pillow==12.1.1
PyJWT==2.11.0
serializers==0.2.4
sqlparse==0.5.5
whitenoise==6.11.0
